srcs/back/game/requirements.txt
Django==4.2.7
channels
channels-redis
bcrypt
django-cors-headers
django-cors-middleware
daphne
psycopg2
djangorestframework
djangorestframework-simplejwt
django-extensions
Werkzeug
django-environ
